+++ crates/backend/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.0.2](https://github.com/napi-rs/napi-rs/compare/napi-derive-backend-v5.0.1...napi-derive-backend-v5.0.2) - 2026-02-14
+
+### Fixed
+
+- *(deps)* update rust crate convert_case to 0.11 ([#3114](https://github.com/napi-rs/napi-rs/pull/3114))
+- *(napi-derive)* escape * in comments ([#3109](https://github.com/napi-rs/napi-rs/pull/3109))
+
 ## [5.0.1](https://github.com/napi-rs/napi-rs/compare/napi-derive-backend-v5.0.0...napi-derive-backend-v5.0.1) - 2026-01-08
 
 ### Fixed